+++ tests/t_bitmap_test.cc
@@ -67,7 +67,7 @@ TEST_F(RedisBitmapTest, BitPosSetBit) {
   }
   int pos;
   int start_indexes[] = {0, 1, 124, 1025, 1027, 3*1024+1};
-  for (int i = 0; i < sizeof(start_indexes)/ sizeof(start_indexes[0]); i++) {
+  for (size_t i = 0; i < sizeof(start_indexes)/ sizeof(start_indexes[0]); i++) {
     bitmap->BitPos(key_, true, start_indexes[i], -1, &pos);
     EXPECT_EQ(pos, offsets[i]);
   }
